refactor(pipeline-email): normalize stages/ + internal/ layout and add canonical pipeline.test.ts
