+++ packages/core/compute/pipeline-email/src/index.ts
@@ -2,5 +2,8 @@
 // Copyright 2026 DXOS.org
 //
 
+// `types/` and `internal/` are intentionally not re-exported here: `types/` is a lower-level
+// building block consumed only within this package, and `internal/` holds helpers with no
+// external consumers (see internal/ modules for the non-re-exported symbols).
 export * from './facts';
 export * from './stages';
