+++ packages/core/compute/pipeline-email/src/internal/threading.ts
