+++ packages/core/compute/pipeline-email/src/stages/extract-facts.ts
