+++ packages/core/compute/pipeline-email/src/stages/index.ts
@@ -1,0 +1,9 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export * from './context';
+export * from './summarize';
+export * from './extract-contacts';
+export * from './stats';
+export * from './extract-facts';
